src/python/requirements.txt
# AI Music Detector - Training Pipeline
# GPU-accelerated feature extraction and model training

# Core ML
torch>=2.0.0
torchaudio>=2.0.0
scikit-learn>=1.3.0

# ONNX export
onnx>=1.20.1
onnxruntime>=1.23.2
onnxruntime-gpu>=1.23.2  # For GPU inference
skl2onnx>=1.19.1

# Data handling
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Audio processing
soxr>=0.3.0
librosa>=0.10.0
soundfile>=0.12.0
nnAudio>=0.3.0  # GPU-accelerated spectrograms

# Visualization
matplotlib>=3.7.0

# Dataset download
huggingface_hub>=0.16.0
requests>=2.31.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
safetensors>=0.4.0
